ml_project/requirements.txt
# local package
-e .

# external requirements
antlr4-python3-runtime 4.8
argon2-cffi            20.1.0
arrow                  1.1.0
astroid                2.5.6
async-generator        1.10
atomicwrites           1.4.0
attrs                  21.2.0
backcall               0.2.0
binaryornot            0.4.4
bleach                 3.3.0
certifi                2020.12.5
cffi                   1.14.5
chardet                4.0.0
click                  7.1.2
colorama               0.4.4
cookiecutter           1.7.2
coverage               5.5
cycler                 0.10.0
decorator              5.0.7
defusedxml             0.7.1
entrypoints            0.3
Faker                  8.1.2
hydra-core             1.0.6
idna                   2.10
importlib-resources    5.1.2
iniconfig              1.1.1
ipykernel              5.5.4
ipython                7.23.1
ipython-genutils       0.2.0
ipywidgets             7.6.3
isort                  5.8.0
jedi                   0.18.0
Jinja2                 2.11.3
jinja2-time            0.2.0
joblib                 1.0.1
jsonschema             3.2.0
jupyter                1.0.0
jupyter-client         6.1.12
jupyter-console        6.4.0
jupyter-core           4.7.1
jupyterlab-pygments    0.1.2
jupyterlab-widgets     1.0.0
kiwisolver             1.3.1
lazy-object-proxy      1.6.0
MarkupSafe             1.1.1
marshmallow            3.12.1
marshmallow-dataclass  8.4.1
matplotlib             3.4.2
matplotlib-inline      0.1.2
mccabe                 0.6.1
mistune                0.8.4
mypy-extensions        0.4.3
nbclient               0.5.3
nbconvert              6.0.7
nbformat               5.1.3
nest-asyncio           1.5.1
notebook               6.3.0
numpy                  1.20.2
omegaconf              2.0.6
packaging              20.9
pandas                 1.2.4
pandocfilters          1.4.3
parso                  0.8.2
pickleshare            0.7.5
Pillow                 8.2.0
pip                    21.1.1
pluggy                 0.13.1
poyo                   0.5.0
prometheus-client      0.10.1
prompt-toolkit         3.0.18
py                     1.10.0
pycparser              2.20
Pygments               2.9.0
pylint                 2.8.2
pyparsing              2.4.7
pyrsistent             0.17.3
pytest                 6.2.4
pytest-cov             2.11.1
python-dateutil        2.8.1
python-slugify         5.0.2
pytz                   2021.1
pywin32                300
pywinpty               0.5.7
PyYAML                 5.4.1
pyzmq                  22.0.3
qtconsole              5.1.0
QtPy                   1.9.0
requests               2.25.1
scikit-learn           0.24.2
scipy                  1.6.3
seaborn                0.11.1
Send2Trash             1.5.0
setuptools             56.1.0
six                    1.16.0
sklearn                0.0
terminado              0.9.4
testpath               0.4.4
text-unidecode         1.3
threadpoolctl          2.1.0
toml                   0.10.2
tornado                6.1
traitlets              5.0.5
typing-extensions      3.10.0.0
typing-inspect         0.7.1
urllib3                1.26.4
wcwidth                0.2.5
webencodings           0.5.1
widgetsnbextension     3.5.1
wrapt                  1.12.1


